requirements.txt
# Lista todas las dependencias
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0